added git sub-modules
The code bases I will be analyzing for my project consist of the
C benchmarking code provided in Professor Fricke's Github repository.

I will also be looking at, and playing with, code from a second
repository related to my hypothesis and experiment. This code
is referenced in the blog post found here:

https://martin-thoma.com/matrix-multiplication-python-java-cpp/

And I will be attempting to replicate the results.