added skeleton code for C benchmarking program
This code is stripped down from the benchmark provided to us for
use in this project. I am going to edit this benchmarking code
to be able to provide an apples to apples comparison to the Java
benchmarking program.

In short: I want to convert from the Mflops metric to just timed
runs using a 100 x 100 matrix multiplication.